+++ src/apiConfig.js
@@ -1,13 +1,13 @@
 let apiUrl;
 const apiUrls = {
-  production: "https://secret-fortress-08155.herokuapp.com",
-  development: "http://localhost:4741",
+    production: "https://secret-fortress-08155.herokuapp.com",
+    development: "http://localhost:4741",
 };
 
-if (window.location.hostname === "localhost") {
-  apiUrl = apiUrls.development;
-} else {
-  apiUrl = apiUrls.production;
-}
+// if (window.location.hostname === "localhost") {
+//   apiUrl = apiUrls.development;
+// } else {
+apiUrl = apiUrls.production;
+// }
 
 export default apiUrl;
