+++ src/components/Maps/ReactMapGL.jsx
@@ -54,67 +54,73 @@ const ReactMapGL = ({
     }, [borough]);
 
     return (
-        <div style={{ display: 'flex', justifyContent: 'center', padding: '10px' }}>
-        {mapData.length > 0 ? (
-            <MapBox>
-                <MapGL
-                    ref={mapRef}
-                    {...viewport}
-                    mapboxApiAccessToken={process.env.REACT_APP_MAPBOX_TOKEN}
-                    onViewportChange={(viewport) => { setViewport(viewport) }}
-                    mapStyle="mapbox://styles/taaseen71/ckleb8llf0zv817lk5y1asq7s"
-                >
-                    {mapData.map((pin) => {
-                        return (
-                            <Marker classname="pin" key={pin._id} latitude={pin.latitude} longitude={pin.longitude} >
-                                <div>
-                                    <DroppedPin
-                                        onClick={(e) => {
-                                            e.preventDefault();
-                                            setPin(pin);
-                                            getRandomImage()
-                                        }}>
-                                        <DroppedPinImage src="./icons/map-icon.svg" alt="Marker Icon" />
-                                    </DroppedPin>
-                                </div>
-                            </Marker>
-                        )
-                    })}
-                    <Geocoder position="top-left" mapRef={mapRef} mapboxApiAccessToken={process.env.REACT_APP_MAPBOX_TOKEN} />
-                    <GeolocateControl
-                        style={{ right: 10, top: 10, zoom: 1 }}
-                        positionOptions={{ enableHighAccuracy: true }}
-                        trackUserLocation={true}
-                        fitBoundsOptions={{ maxZoom: 10 }}
-                        showAccuracyCircle={true}
-                    />
-                    	{pin && (
-					<Popup
-						user={user}
-						closeOnClick={false}
-						closeButton={true}
-						latitude={pin.latitude}
-						longitude={pin.longitude}
-						onClose={() => handleClose()}
-                        alert={alert}
-                        >
-                        <LocationCardTwo
-                            pin={pin}
-                            user={user}
-                            alert={alert}
-                            randomImage={randomImage} />
-                    </Popup>
-                        )}  
-                </MapGL>
-            </MapBox>
-             ) : (
-            <LoadingWrapper>
-                <Spinner className="center" animation="grow" variant="success" />
-            </LoadingWrapper>
-            )}
-        </div>
+        <MapDiv >
+            {mapData.length > 0 ? (
+                <MapBox>
+                    <MapGL
+                        ref={mapRef}
+                        {...viewport}
+                        mapboxApiAccessToken={process.env.REACT_APP_MAPBOX_TOKEN}
+                        onViewportChange={(viewport) => { setViewport(viewport) }}
+                        mapStyle="mapbox://styles/taaseen71/ckleb8llf0zv817lk5y1asq7s"
+                    >
+                        {mapData.map((pin) => {
+                            return (
+                                <Marker classname="pin" key={pin._id} latitude={pin.latitude} longitude={pin.longitude} >
+                                    <div>
+                                        <DroppedPin
+                                            onClick={(e) => {
+                                                e.preventDefault();
+                                                setPin(pin);
+                                                getRandomImage()
+                                            }}>
+                                            <DroppedPinImage src="./icons/Marker-Icon.svg" alt="Marker Icon" />
+                                        </DroppedPin>
+                                    </div>
+                                </Marker>
+                            )
+                        })}
+                        <Geocoder position="top-left" mapRef={mapRef} mapboxApiAccessToken={process.env.REACT_APP_MAPBOX_TOKEN} />
+                        <GeolocateControl
+                            style={{ right: 10, top: 10, zoom: 1 }}
+                            positionOptions={{ enableHighAccuracy: true }}
+                            trackUserLocation={true}
+                            fitBoundsOptions={{ maxZoom: 10 }}
+                            showAccuracyCircle={true}
+                        />
+                        {pin && (
+                            <Popup
+                                user={user}
+                                closeOnClick={false}
+                                closeButton={true}
+                                latitude={pin.latitude}
+                                longitude={pin.longitude}
+                                onClose={() => handleClose()}
+                                alert={alert}
+                            >
+                                <LocationCardTwo
+                                    pin={pin}
+                                    user={user}
+                                    alert={alert}
+                                    randomImage={randomImage} />
+                            </Popup>
+                        )}
+                    </MapGL>
+                </MapBox>
+            ) : (
+                    <LoadingWrapper>
+                        <Spinner className="center" animation="grow" variant="success" />
+                    </LoadingWrapper>
+                )}
+        </MapDiv>
     )
 }
+
+const MapDiv = styled.div`
+    display: flex;
+    justify-content: center;
+    padding: 10px;
+`
 
 const LoadingWrapper = styled.div`
   display: flex;
